+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed Explicit minor versioning of dependencies [#3](https://github.com/as1100k/bevy-discord/pull/3)
 - Disabled default features for `reqwest` [#3](https://github.com/as1100k/bevy-discord/pull/3)
 - Improved Documentation
+- Moved `bevy_discord::bot::serenity` to `bevy_discord::serenity`
 
 ### Deprecated
 - module `webhook`